Blast phases now adapted to the current framerate.  New constant for speed of animation independant of framerate.

+++ src/struct.h
@@ -144,7 +144,7 @@ typedef struct {
    int PX;  /* PosX */
    int PY;  /* PosY */
    byte type;
-   byte phase;
+   float phase;
 } blast, *Blast;
 
 typedef struct {
@@ -222,6 +222,9 @@ typedef struct {
  * $Author$
  *
  * $Log$
+ * Revision 1.20  1997/06/09 23:08:59  jprix
+ * Blast phases now adapted to the current framerate.  New constant for speed of animation independant of framerate.
+ *
  * Revision 1.19  1997/06/09 21:53:49  jprix
  * Rotation of enemys and influencer now independant of the framerate.
  *
